backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Korean Learning App Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "python-dotenv>=0.19.0",
    "sqlmodel>=0.0.8",
    "groq>=0.4.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]
